feat: add CSSWrapper tests
