+++ src/utilities/css-wrapper/CSSWrapper.tsx
@@ -1,0 +1,21 @@
+import * as React from 'react'
+
+import { Box } from '../../primitives'
+import { CSSBlob } from '../../stitches'
+
+type CssWrapperProps = {
+  css?: CSSBlob
+  children: React.ReactNode
+}
+
+export const CSSWrapper = ({
+  css,
+  children
+}: CssWrapperProps): React.ReactElement =>
+  css ? (
+    <Box css={css} data-testid="css-wrapper">
+      {children}
+    </Box>
+  ) : (
+    <> children </>
+  )
