+++ src/utilities/css-wrapper/index.ts
@@ -1,0 +1,1 @@
+export { CSSWrapper } from './CSSWrapper'
